backend/requirements.txt
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.21.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.27.0
psycopg2-binary
pydantic[email]>=2.0.0,<3.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0,<4.1.0
passlib[bcrypt]>=1.7.4
python-multipart
minio>=7.2.15
python-dotenv>=1.0.0
alembic>=1.9.0
fastapi-mail>=1.2.0
greenlet>=2.0.0
slowapi>=0.1.0
websockets>=10.4
httpx>=0.24.0
